+++ Inventory/src/main/java/com/bartolay/inventory/repositories/LocationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.bartolay.inventory.entity.Location;
 
 @Repository
-public interface LocationRepository extends CrudRepository<Location, Long> {
+public interface LocationRepository extends CrudRepository<Location, Integer> {
 
 }
